Create Selection sort.cpp
In the Selection sort sorting algorithm, we iterate over each element in the array and we find the minimum element starting from the next element of the element that is being pointed right now, and we store the index of that element. Coming out of the inner for loop, we swap the element currently being iterated and the element at the index. In this way the with each iteration the array gets sorted starting from the first element.
This algorithm takes O(n^2) time complexity.