Create Contiguous-Array.cpp
In this question, I used Hashmap to store the prefix sums of the indices and using them, we will check if this sum already exists in the Hashmap, if it does then we will update the subArray_len by comparing the current length with the previous one. Else we will add a new prefix sum with its index in the Hashmap.
Prefix sums are calculated by incrementing the sum variable by 1 on finding 1 and decrementing by 1 on finding 0.